+++ CXPMobileSDK/mfe3e4a-login/DemoWidgetsAssets/assets/backbase/static/features/[BBHOST]/module-charts/scripts/components/line-chart/scripts/line-chart.js
@@ -9,6 +9,7 @@ define(function(require, exports, module) {
     var tooltipFactory = require('../../../libs/tooltip');
     var lineFactory = require('../../../libs/line');
     var pointsFactory = require('../../../libs/points');
+    var isSVGAvailable = require('../../../utils').isSVGAvailable();
 
     function getTicks(data) {
         // If the data are 7 days print them all, otherwise it is a month so take data
@@ -54,8 +55,17 @@ define(function(require, exports, module) {
     exports.lpLineChart = function($window){
         function link(scope, element) {
             var $canvas = element.find('.canvas');
-            var svg = d3.select(element.find('svg')[0]);
-            var graph = svg.append('g');
+            var svg, graph;
+
+            if (isSVGAvailable) {
+                svg = d3.select(element.find('svg')[0]);
+                graph = svg.append('g');
+            } else {
+                $canvas.empty();
+                svg = d3.select($canvas[0]).append('svg:svg');
+                svg.attr('class', 'chart');
+                graph = svg.append('svg:g');
+            }
 
             var x = d3.time.scale();
             var y = d3.scale.linear();
@@ -70,7 +80,7 @@ define(function(require, exports, module) {
             var axes = axesFactory(config);
             var line = lineFactory(config);
             var points = pointsFactory(pointsConfig).attr('class', 'points');
-            var tooltip = tooltipFactory(config);
+            var tooltip = isSVGAvailable ? tooltipFactory(config) : null;
 
 
             function render() {
@@ -95,7 +105,7 @@ define(function(require, exports, module) {
                 y.range([height, 0]);
 
                 axes.resize(width, height);
-                tooltip.resize(width, height);
+                if (isSVGAvailable) { tooltip.resize(width, height); }
 
                 render();
             }