+++ CXPMobileSDK/mfe3e4a-login/DemoWidgetsAssets/assets/backbase/static/widgets/[BBHOST]/widget-profile-contact/scripts/main.js
@@ -6,9 +6,12 @@ define(function(require, exports, module) {
     var ui = require('ui');
     var core = require('core');
 
+    var users = require('module-users');
+
     var deps = [
         core.name,
-        ui.name
+        ui.name,
+        users.name
     ];
 
     // @ngInject