+++ CXPMobileSDK/mfe3e4b-logout/DemoWidgetsAssets/assets/backbase/static/features/[BBHOST]/base/scripts/modules/session/session-handler.spec.js
@@ -19,6 +19,29 @@ describe('Session handler', function() {
     it('should export a function', function() {
         expect(Session.getInstance).toBeFunction();
     });
+    describe('Session handler instance', function() {
+        var sessionHandler = Session.getInstance();
+        console.log(sessionHandler);
+        it('should export an object', function() {
+            expect(sessionHandler).toBeObject();
+        });
+        it('should be a function', function() {
+            expect(sessionHandler.init).toBeFunction();
+            expect(sessionHandler.logout).toBeFunction();
+            expect(sessionHandler.startPolling).toBeFunction();
+            expect(sessionHandler.ping).toBeFunction();
+            expect(sessionHandler.makeSessionRequest).toBeFunction();
+            expect(sessionHandler.handleStateResponse).toBeFunction();
+            expect(sessionHandler.logError).toBeFunction();
+            expect(sessionHandler.handleNetworkError).toBeFunction();
+            expect(sessionHandler.startSessionWarning).toBeFunction();
+            expect(sessionHandler.clearSessionWarning).toBeFunction();
+        });
+        it('should be false', function() {
+            expect(sessionHandler.hasSession).toBeFalse();
+            expect(sessionHandler.isPollingStarted).toBeFalse();
+        });
+    });
 });